Add check

+++ python/cudaq/kernel/kernel_decorator.py
@@ -550,7 +550,8 @@ class PyKernelDecorator(object):
         Delete the `ExecutionEngine` cache given by a cache key.
         """
         # Make sure this hasn't already been cleaned up as we're winding down
-        if cudaq_runtime.delete_cache_execution_engine is not None:
+        if (cudaq_runtime is not None and
+            cudaq_runtime.delete_cache_execution_engine is not None):
             cudaq_runtime.delete_cache_execution_engine(key)
 
     def resolve_decorator_at_callsite(self, callingMod):
